+++ client/lib/ft4/accounts/transfer-history/transfer-history-queries.ts
@@ -1,4 +1,4 @@
-import { OptionalPageCursor } from "@ft4/ft-session";
+import { OptionalPageCursor, PagedResponse } from "@ft4/ft-session";
 import { Buffer } from "buffer";
 import { BufferId, QueryObject, formatter } from "postchain-client";
 import { RawTransferDetail, TransferHistoryEntryResponse } from "./types";
@@ -65,7 +65,7 @@ export function transferHistoryFromHeight(
   limit: number,
   cursor: OptionalPageCursor = null,
 ): QueryObject<
-  TransferHistoryEntryResponse[],
+  PagedResponse<TransferHistoryEntryResponse>,
   {
     height: number;
     asset_id: Buffer | null;