+++ client/lib/ft4/registration/main.ts
@@ -11,10 +11,12 @@ import {
 import { compactArray, createAndSignTransaction } from "@ft4/utils";
 import {
   IClient,
+  MERKLE_HASH_VERSIONS,
   Operation,
   Queryable,
-  ResponseStatus,
   SignedTransaction,
+  TransactionEvent,
+  Web3PromiEvent,
   gtv,
 } from "postchain-client";
 import { registerAccount as registerAccountOp } from "./operations";
@@ -27,7 +29,6 @@ import {
   createSession,
 } from "@ft4/ft-session";
 import { evmSignatures } from "@ft4/transaction-builder/utils";
-import { Web3CustomPromiEvent } from "@ft4/utils/promiEvent";
 
 /**
  * Registers an account.
@@ -40,14 +41,14 @@ export function registerAccount(
   masterKeyStore: FtKeyStore | EvmKeyStore,
   strategy: Strategy,
   registerAccountOperation: Operation = registerAccountOp(),
-): Web3CustomPromiEvent<
+): Web3PromiEvent<
   SessionWithLogout,
   {
     built: SignedTransaction;
     sent: Buffer;
   }
 > {
-  const promiEvent = new Web3CustomPromiEvent<
+  const promiEvent = new Web3PromiEvent<
     SessionWithLogout,
     {
       built: SignedTransaction;
@@ -119,15 +120,16 @@ export function registerAccount(
           keyStores,
           connection.client
             .sendTransaction(transaction)
-            .on("sent", (receipt) => {
-              if (receipt.status === ResponseStatus.Waiting) {
-                promiEvent.emit("sent", receipt.transactionRid);
-              }
+            .on(TransactionEvent.DappReceived, (receipt) => {
+              promiEvent.emit("sent", receipt.transactionRid);
             }),
         ]);
       })
       .then(([loginKeyStore, disposableKeyStore, keyStores, _]) => {
-        const accountId = gtv.gtvHash(masterKeyStore.id);
+        const accountId = gtv.gtvHash(
+          masterKeyStore.id,
+          MERKLE_HASH_VERSIONS.ONE,
+        );
 
         return Promise.all([
           loginKeyStore,