+++ client/lib/ft4/utils/index.ts
@@ -41,11 +41,3 @@ export {
   getBlockchainApiUrls,
   getDirectoryClient,
 } from "./directory-chain";
-
-export {
-  Web3CustomPromiEvent,
-  Web3CustomEventMap,
-  PromiseExecutor,
-  Web3EventCallback,
-  Web3EventKey,
-} from "./promiEvent";