+++ scripts/update-docs.sh
@@ -11,7 +11,7 @@ npx typedoc --out docs-repo/static/pages/ft4-ts-client/client
 
 # create rell docs
 chr install
-chr build
+chr build --hide-lib-warnings
 chr generate docs-site --target docs-repo/static/pages/ft4-rell -s doc/docs-chromia.yml
 
 # create changelogs