+++ test/integration/assets/asset.integration.test.ts
@@ -11,7 +11,12 @@ import { Connection, createConnection } from "@ft4/ft-session";
 import { op } from "@ft4/utils";
 import { Buffer } from "buffer";
 import { randomBytes } from "crypto";
-import { IClient, formatter, gtv } from "postchain-client";
+import {
+  IClient,
+  MERKLE_HASH_VERSIONS,
+  formatter,
+  gtv,
+} from "postchain-client";
 
 let connection: Connection;
 let client: IClient;
@@ -92,7 +97,10 @@ describe("Asset", () => {
       connection.client.config.blockchainRid,
       "hex",
     );
-    const assetId = gtv.gtvHash([assetName, blockchainRid]);
+    const assetId = gtv.gtvHash(
+      [assetName, blockchainRid],
+      MERKLE_HASH_VERSIONS.ONE,
+    );
     await getNewAsset(client, assetName, assetSymbol, 3);
 
     const expectedAsset = await connection.getAssetById(assetId);
@@ -106,7 +114,10 @@ describe("Asset", () => {
   it("returns an asset when queried by symbol", async () => {
     const assetName = "asset_symbol";
     const assetSymbol = "ASSET_SYMBOL";
-    const assetId = gtv.gtvHash([assetName, blockchainRid]);
+    const assetId = gtv.gtvHash(
+      [assetName, blockchainRid],
+      MERKLE_HASH_VERSIONS.ONE,
+    );
     const iconUrl = "http://example.com/";
     await getNewAsset(client, assetName, assetSymbol, 3, iconUrl);
 
@@ -117,10 +128,10 @@ describe("Asset", () => {
     const crosschainBlockchainRid = formatter.ensureBuffer("34".repeat(32));
     const crosschainIconUrl = "";
     const crosschainAssetType = "FT4";
-    const crosschainRes = gtv.gtvHash([
-      crosschainAssetName,
-      crosschainBlockchainRid,
-    ]);
+    const crosschainRes = gtv.gtvHash(
+      [crosschainAssetName, crosschainBlockchainRid],
+      MERKLE_HASH_VERSIONS.ONE,
+    );
 
     await registerCrosschainAsset(
       connection,