+++ test/unit/pathfinder.test.ts
@@ -35,6 +35,7 @@ import {
   MissingNodeUrlError,
   formatter,
   BufferId,
+  MERKLE_HASH_VERSIONS,
 } from "postchain-client";
 import { Connection, createConnection } from "@ft4/ft-session";
 import { Asset } from "@ft4/asset/types";
@@ -58,6 +59,7 @@ describe("Pathfinder", () => {
     const client = await createClient({
       nodeUrlPool: "",
       blockchainRid: formatter.toString(startingChainRid),
+      merkleHashVersion: MERKLE_HASH_VERSIONS.ONE,
     });
     connection = createConnection(client);
   });