+++ test/util/test-user.ts
@@ -12,6 +12,7 @@ import {
   gtv,
   gtx,
   KeyPair,
+  MERKLE_HASH_VERSIONS,
   SignatureProvider,
 } from "postchain-client";
 
@@ -25,7 +26,10 @@ export function newSingleSigUser(
   keyPair: KeyPair,
   rule: AuthDescriptorRules | null = null,
 ): User {
-  const signatureProvider = gtx.newSignatureProvider(keyPair);
+  const signatureProvider = gtx.newSignatureProvider(
+    MERKLE_HASH_VERSIONS.ONE,
+    keyPair,
+  );
   const singleSigAuthDescriptor = createSingleSigAuthDescriptorRegistration(
     [AuthFlag.Account, AuthFlag.Transfer],
     signatureProvider.pubKey,
@@ -36,7 +40,7 @@ export function newSingleSigUser(
     authDescriptor: {
       ...singleSigAuthDescriptor,
       id: deriveAuthDescriptorId(singleSigAuthDescriptor),
-      accountId: gtv.gtvHash(keyPair.pubKey),
+      accountId: gtv.gtvHash(keyPair.pubKey, MERKLE_HASH_VERSIONS.ONE),
       accountType: FT4_USER_TYPE,
       created: new Date(),
     },